Moved hooks into a separate directory
