+++ gatsby-theme-shopify-core/src/hooks/__tests__/.eslintrc
@@ -1,0 +1,6 @@
+{
+  "rules": {
+    "@typescript-eslint/ban-ts-ignore": "off",
+    "@typescript-eslint/no-non-null-assertion": "off"
+  }
+}
