+++ gatsby-theme-shopify-core/src/hooks/__tests__/useAddItemToCart.test.tsx
@@ -1,16 +1,28 @@
 import React, {useState} from 'react';
 import {wait, fireEvent} from '@testing-library/react';
-import {renderWithContext} from '../mocks';
-import {LocalStorage, LocalStorageKeys} from '../utils';
-import {useAddItemsToCart} from '../useAddItemsToCart';
-import {LineItemPatch} from '../types';
+import {renderWithContext} from '../../mocks';
+import {LocalStorage, LocalStorageKeys} from '../../utils';
+import {useAddItemToCart} from '../useAddItemToCart';
+import {AttributeInput} from '../../types';
 
-function MockComponent({items}: {items: LineItemPatch[]}) {
-  const addItemsToCart = useAddItemsToCart();
+function MockComponent({
+  variantId,
+  quantity,
+  customAttributes,
+}: {
+  variantId: string | number;
+  quantity: number;
+  customAttributes?: AttributeInput[];
+}) {
+  const addItemToCart = useAddItemToCart();
   const [result, setResult] = useState<boolean | null>(null);
 
   async function addItem() {
-    const newResult = await addItemsToCart(items);
+    const newResult = await addItemToCart(
+      variantId,
+      quantity,
+      customAttributes,
+    );
     setResult(newResult);
   }
 
@@ -37,17 +49,10 @@ afterEach(() => {
   console.error = originalError;
 });
 
-describe('useAddItemsToCart()', () => {
-  it('returns true if the items are added to the cart', async () => {
+describe('useAddItemToCart()', () => {
+  it('returns true if the item is added to the cart', async () => {
     const wrapper = renderWithContext(
-      <MockComponent
-        items={[
-          {
-            variantId: 'variantId',
-            quantity: 1,
-          },
-        ]}
-      />,
+      <MockComponent variantId={'some_variant_id'} quantity={1} />,
     );
     await wait(() => {
       fireEvent.click(wrapper.getByText(/Add to Cart/));
@@ -59,14 +64,7 @@ describe('useAddItemsToCart()', () => {
   it('updates the cart state if the items are added to the cart', async () => {
     const localStorageSpy = jest.spyOn(LocalStorage, 'set');
     const wrapper = renderWithContext(
-      <MockComponent
-        items={[
-          {
-            variantId: 'newVariantId',
-            quantity: 1,
-          },
-        ]}
-      />,
+      <MockComponent variantId="newVariantId" quantity={1} />,
     );
     await wait(() => {
       fireEvent.click(wrapper.getByText(/Add to Cart/));
@@ -81,26 +79,19 @@ describe('useAddItemsToCart()', () => {
     expect(localStorageSpy).toHaveBeenCalledTimes(3);
   });
 
-  it('returns false if there are no items passed to the function', async () => {
-    const wrapper = renderWithContext(<MockComponent items={[]} />);
+  it('returns false if there is no variantId passed to the function', async () => {
+    // @ts-ignore
+    const wrapper = renderWithContext(<MockComponent quantity={1} />);
     await wait(() => {
       fireEvent.click(wrapper.getByText(/Add to Cart/));
     });
 
     expect(wrapper.getByText(/Result: false/)).toBeTruthy();
-    expect(mockConsoleError).toHaveBeenCalledWith(
-      'Must include at least one line item, empty line items found',
-    );
   });
 
-  it("returns false if the input does not conform to the client's protocol", async () => {
-    const items = [{variantId: 'variantId'}];
-    const wrapper = renderWithContext(
-      <MockComponent
-        // @ts-ignore
-        items={items}
-      />,
-    );
+  it('returns false if there is no quantity passed to the function', async () => {
+    // @ts-ignore
+    const wrapper = renderWithContext(<MockComponent variantId="someId" />);
     await wait(() => {
       fireEvent.click(wrapper.getByText(/Add to Cart/));
     });
