+++ gatsby-theme-shopify-core/src/hooks/__tests__/useAddItemsToCart.test.tsx
@@ -1,28 +1,16 @@
 import React, {useState} from 'react';
 import {wait, fireEvent} from '@testing-library/react';
-import {renderWithContext} from '../mocks';
-import {LocalStorage, LocalStorageKeys} from '../utils';
-import {useAddItemToCart} from '../useAddItemToCart';
-import {AttributeInput} from '../types';
+import {renderWithContext} from '../../mocks';
+import {LocalStorage, LocalStorageKeys} from '../../utils';
+import {useAddItemsToCart} from '../useAddItemsToCart';
+import {LineItemPatch} from '../../types';
 
-function MockComponent({
-  variantId,
-  quantity,
-  customAttributes,
-}: {
-  variantId: string | number;
-  quantity: number;
-  customAttributes?: AttributeInput[];
-}) {
-  const addItemToCart = useAddItemToCart();
+function MockComponent({items}: {items: LineItemPatch[]}) {
+  const addItemsToCart = useAddItemsToCart();
   const [result, setResult] = useState<boolean | null>(null);
 
   async function addItem() {
-    const newResult = await addItemToCart(
-      variantId,
-      quantity,
-      customAttributes,
-    );
+    const newResult = await addItemsToCart(items);
     setResult(newResult);
   }
 
@@ -49,10 +37,17 @@ afterEach(() => {
   console.error = originalError;
 });
 
-describe('useAddItemToCart()', () => {
-  it('returns true if the item is added to the cart', async () => {
+describe('useAddItemsToCart()', () => {
+  it('returns true if the items are added to the cart', async () => {
     const wrapper = renderWithContext(
-      <MockComponent variantId={'some_variant_id'} quantity={1} />,
+      <MockComponent
+        items={[
+          {
+            variantId: 'variantId',
+            quantity: 1,
+          },
+        ]}
+      />,
     );
     await wait(() => {
       fireEvent.click(wrapper.getByText(/Add to Cart/));
@@ -64,7 +59,14 @@ describe('useAddItemToCart()', () => {
   it('updates the cart state if the items are added to the cart', async () => {
     const localStorageSpy = jest.spyOn(LocalStorage, 'set');
     const wrapper = renderWithContext(
-      <MockComponent variantId="newVariantId" quantity={1} />,
+      <MockComponent
+        items={[
+          {
+            variantId: 'newVariantId',
+            quantity: 1,
+          },
+        ]}
+      />,
     );
     await wait(() => {
       fireEvent.click(wrapper.getByText(/Add to Cart/));
@@ -79,19 +81,26 @@ describe('useAddItemToCart()', () => {
     expect(localStorageSpy).toHaveBeenCalledTimes(3);
   });
 
-  it('returns false if there is no variantId passed to the function', async () => {
-    // @ts-ignore
-    const wrapper = renderWithContext(<MockComponent quantity={1} />);
+  it('returns false if there are no items passed to the function', async () => {
+    const wrapper = renderWithContext(<MockComponent items={[]} />);
     await wait(() => {
       fireEvent.click(wrapper.getByText(/Add to Cart/));
     });
 
     expect(wrapper.getByText(/Result: false/)).toBeTruthy();
+    expect(mockConsoleError).toHaveBeenCalledWith(
+      'Must include at least one line item, empty line items found',
+    );
   });
 
-  it('returns false if there is no quantity passed to the function', async () => {
-    // @ts-ignore
-    const wrapper = renderWithContext(<MockComponent variantId="someId" />);
+  it("returns false if the input does not conform to the client's protocol", async () => {
+    const items = [{variantId: 'variantId'}];
+    const wrapper = renderWithContext(
+      <MockComponent
+        // @ts-ignore
+        items={items}
+      />,
+    );
     await wait(() => {
       fireEvent.click(wrapper.getByText(/Add to Cart/));
     });
