+++ gatsby-theme-shopify-core/src/hooks/__tests__/useCartItems.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {wait} from '@testing-library/react';
-import {Mocks, renderWithContext} from '../mocks';
-import {LocalStorage, LocalStorageKeys} from '../utils';
+import {Mocks, renderWithContext} from '../../mocks';
+import {LocalStorage, LocalStorageKeys} from '../../utils';
 import {useCartItems} from '../useCartItems';
 
 function MockComponent() {
