+++ gatsby-theme-shopify-core/src/hooks/useAddItemsToCart.tsx
@@ -1,7 +1,7 @@
 import {useContext} from 'react';
-import {Context} from './Context';
+import {Context} from '../Context';
 import ShopifyBuy from 'shopify-buy';
-import {LineItemPatch} from './types';
+import {LineItemPatch} from '../types';
 
 export function useAddItemsToCart() {
   const {client, cart, setCart} = useContext(Context);
