+++ gatsby-theme-shopify-core/src/hooks/useCart.tsx
@@ -1,5 +1,5 @@
 import {useContext} from 'react';
-import {Context} from './Context';
+import {Context} from '../Context';
 
 export function useCart() {
   const {cart, setCart} = useContext(Context);
