+++ gatsby-theme-shopify-core/src/hooks/useCartItems.tsx
@@ -1,5 +1,5 @@
 import {useContext} from 'react';
-import {Context} from './Context';
+import {Context} from '../Context';
 
 export function useCartItems() {
   const {cart} = useContext(Context);
