+++ gatsby-theme-shopify-core/src/hooks/useUpdateItemQuantity.tsx
@@ -1,5 +1,5 @@
 import {useContext} from 'react';
-import {Context} from './Context';
+import {Context} from '../Context';
 
 import {useGetLineItem} from './useGetLineItem';
