+++ gatsby-theme-shopify-core/src/index.ts
@@ -1,10 +1,2 @@
 export {ContextProvider} from './ContextProvider';
-export {useClient} from './useClient';
-export {useCart} from './useCart';
-export {useCartCount} from './useCartCount';
-export {useAddItemToCart} from './useAddItemToCart';
-export {useAddItemsToCart} from './useAddItemsToCart';
-export {useCartItems} from './useCartItems';
-export {useCheckoutUrl} from './useCheckoutUrl';
-export {useGetLineItem} from './useGetLineItem';
-export {useUpdateItemQuantity} from './useUpdateItemQuantity';
+export * from './hooks';
